Core: Add functions for pakcing and unpacking pixels

+++ src/gba/gba.c
@@ -856,7 +856,7 @@ void GBAFrameEnded(struct GBA* gba) {
 
 	if (gba->stream && gba->stream->postVideoFrame) {
 		const color_t* pixels;
-		unsigned stride;
+		size_t stride;
 		gba->video.renderer->getPixels(gba->video.renderer, &stride, (const void**) &pixels);
 		gba->stream->postVideoFrame(gba->stream, pixels, stride);
 	}
